Added small buttons and other stuff
I added some small buttons in the bottom for all kinds of stuff... Don't do anything, but that's a smaller issue :o

Fixed something that was bad

Made a decent redesign, it's lookin real nice, and it adapts pretty well in light and dark mode. Which is nice :3! Also it just looks nice in general.