Added a bunch of shit.
Some of that shit being the ability to toggle the Apple Music Logo... A little bit better randomizer, still need improvements but that ain't happening right now, just because I say so.

Not much other than that, however it feels like I added a bunch of other shit, but I guess not.