Fixed issue #1
The problem apparently was a lot different then what I original thought was the issue.
But after checking a few things I found out out why this occurs.

The reason this issue even happens is when a filepath has a space in it, I never tried a file with such and so I never got to experience this issue first hand :p